pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42", # At least v42 of setuptools required.
  "versioningit",
]

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.0"

[project]
name = "meltano-tap-cratedb"
description = "A Singer tap / Meltano extractor for CrateDB, built with the Meltano SDK, and based on the Meltano PostgreSQL tap."
readme = "README.md"
keywords = [
  "CrateDB",
  "data-loading",
  "data-processing",
  "data-toolkit",
  "data-transfer",
  "data-transformation",
  "ELT",
  "ETL",
  "io",
  "Meltano",
  "Meltano SDK",
  "PostgreSQL",
  "Singer",
]
license = { text = "MIT" }
authors = [
  { name = "Andreas Motl", email = "andreas.motl@crate.io" },
]
requires-python = ">=3.8.1,<3.12"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Customer Service",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Manufacturing",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Telecommunications Industry",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Unix",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Adaptive Technologies",
  "Topic :: Communications",
  "Topic :: Database",
  "Topic :: Documentation",
  "Topic :: Education",
  "Topic :: Internet",
  "Topic :: Office/Business",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Pre-processors",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
  "Topic :: Software Development :: Version Control",
  "Topic :: System :: Archiving",
  "Topic :: System :: Benchmark",
  "Topic :: System :: Clustering",
  "Topic :: System :: Distributed Computing",
  "Topic :: System :: Logging",
  "Topic :: System :: Monitoring",
  "Topic :: System :: Networking",
  "Topic :: System :: Systems Administration",
  "Topic :: Text Processing",
  "Topic :: Utilities",
]
dynamic = [
  "version",
]
dependencies = [
  "crate[sqlalchemy]",
  "cratedb-toolkit",
  'importlib-resources; python_version < "3.9"',
  "meltanolabs-tap-postgres==0.0.6",
]
[project.optional-dependencies]
develop = [
  "black<24",
  "mypy==1.7.1",
  "poethepoet<0.25",
  "pyproject-fmt<1.6",
  "ruff==0.1.7",
  "validate-pyproject<0.16",
]
release = [
  "build<2",
  "twine<5",
]
test = [
  "pytest<8",
  "pytest-cov<5",
  "pytest-mock<4",
]
[project.urls]
changelog = "https://github.com/crate-workbench/meltano-tap-cratedb/blob/main/CHANGES.md"
documentation = "https://github.com/crate-workbench/meltano-tap-cratedb"
homepage = "https://github.com/crate-workbench/meltano-tap-cratedb"
repository = "https://github.com/crate-workbench/meltano-tap-cratedb"
[project.scripts]
tap-cratedb = 'tap_cratedb.tap:TapCrateDB.cli'

[tool.black]
line-length = 120

[tool.coverage.run]
branch = false
omit = [
  "tests/*",
]
source = ["tap_cratedb"]

[tool.coverage.report]
fail_under = 0
show_missing = true

[tool.mypy]
packages = ["tap_cratedb"]
exclude = [
]
check_untyped_defs = true
ignore_missing_imports = true
implicit_optional = true
install_types = true
non_interactive = true

[tool.pytest.ini_options]
addopts = """
  -rfEXs -p pytester --strict-markers --verbosity=3
  --cov --cov-report=term-missing --cov-report=xml
  """
minversion = "2.0"
log_level = "DEBUG"
log_cli_level = "DEBUG"
log_format = "%(asctime)-15s [%(name)-36s] %(levelname)-8s: %(message)s"
testpaths = [
  "examples",
  "tap_cratedb",
  "tests",
]
xfail_strict = true
markers = [
]

[tool.ruff]
line-length = 120

select = [
  # Bandit
  "S",
  # Bugbear
  "B",
  # Builtins
  "A",
  # comprehensions
  "C4",
  # eradicate
  "ERA",
  # flake8-2020
  "YTT",
  # isort
  "I",
  # pandas-vet
  "PD",
  # print
  "T20",
  # Pycodestyle
  "E",
  "W",
  # Pyflakes
  "F",
  # return
  "RET",
]

extend-ignore = [
  # zip() without an explicit strict= parameter
  "B905",
  # df is a bad variable name. Be kinder to your future self.
  "PD901",
  # Unnecessary variable assignment before `return` statement
  "RET504",
  # Unnecessary `elif` after `return` statement
  "RET505",
  # Probable insecure usage of temporary file or directory
  "S108",
]

extend-exclude = [
]

[tool.ruff.per-file-ignores]
"*/tests/*" = [
  "S101",  # Allow use of `assert`, and `print`.
  "S608",  # Possible SQL injection vector through string-based query construction.
]
"examples/*" = ["T201"]  # Allow `print`

[tool.setuptools.packages.find]
namespaces = false


# ===================
# Tasks configuration
# ===================

[tool.poe.tasks]

check = [
  "lint",
  "test",
]

format = [
  { cmd = "black ." },
  # Configure Ruff not to auto-fix (remove!):
  # unused imports (F401), unused variables (F841), `print` statements (T201), and commented-out code (ERA001).
  { cmd = "ruff --fix --ignore=ERA --ignore=F401 --ignore=F841 --ignore=T20 --ignore=ERA001 ." },
  { cmd = "pyproject-fmt pyproject.toml" },
]

lint = [
  { cmd = "ruff ." },
  { cmd = "black --check ." },
  { cmd = "validate-pyproject pyproject.toml" },
  { cmd = "mypy" },
]

release = [
  { cmd = "python -m build" },
  { cmd = "twine upload dist/*" },
]

test = { cmd = "pytest" }
